refactor(db): extract reaction repository
